Add version 1.4.1

+++ src/units.cpp
@@ -28,9 +28,9 @@ double Units::length(Point p1, Point p2) const {
 	return length(p1.x, p1.y, p2.x, p2.y);
 }
 
-double Units::length(int x1, int y1, int x2, int y2) const {
-	int dx = m_xResolution * (x1 - x2);
-	int dy = m_yResolution * (y1 - y2);
+double Units::length(double x1, double y1, double x2, double y2) const {
+	double dx = m_xResolution * (x1 - x2);
+	double dy = m_yResolution * (y1 - y2);
 	return std::sqrt(dx * dx + dy * dy);
 }
